% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.adjust.R
\name{calculate.dra.risk.adjustment.glm}
\alias{calculate.dra.risk.adjustment.glm}
\title{Calculates a GLM for using in the survey direct risk adjustment process.}
\usage{
calculate.dra.risk.adjustment.glm(
  input.dt,
  outcome.col.name,
  covariate.col.names,
  outlier.limits = c(-Inf, Inf),
  ...
)
}
\arguments{
\item{input.dt}{Input data.table with all cited columns present.}

\item{outcome.col.name}{The outcome column.}

\item{covariate.col.names}{A character vector of column names to add to the
model.}

\item{outlier.limits}{Vector air of numeric. Outcomes lower than the first or
higher than the second will be excluded from the model. Might be a bit
bizarre for factors (default: c(-Inf, Inf))}

\item{...}{Extra variables to be passed to stats::glm. In particular, you'll
probably want to specify the family (e.g. binomial(link = 'logit') for
binary, poisson for length of stay)}
}
\value{
A GLM object.
}
\description{
Calculates a GLM for using in the survey direct risk adjustment process.
}
