% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.adjust.R
\name{calculate.quantiles}
\alias{calculate.quantiles}
\title{Calculates the quantile for each of a list of numbers, can be provided a list 
of numbers that it will round the percentile to. That should be useful for
using the correct quantile from the quantile regression to correct the
output.variable.}
\usage{
calculate.quantiles(input.vector, possible.values = NULL)
}
\arguments{
\item{input.vector}{Vector of numbers to calculate quantiles for.}

\item{possible.values}{The values that the quantiles will be rounded to.}
}
\value{
Vector of quantiles the same size as input
}
\description{
Calculates the quantile for each of a list of numbers, can be provided a list 
of numbers that it will round the percentile to. That should be useful for
using the correct quantile from the quantile regression to correct the
output.variable.
}
