% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{calculate.summary.dt}
\alias{calculate.summary.dt}
\title{Summarise a data.table}
\usage{
calculate.summary.dt(
  input.dt,
  daoh.col.name = "daoh",
  by.group = NULL,
  weight.col.name = NULL
)
}
\arguments{
\item{daoh.col.name}{Name of the DAOH column}

\item{by.group}{Name of the stratifying group column, none if NA (e.g., an
intervention) (Default: NULL)}

\item{weight.col.name}{Name of a column to weight observations by, for
risk-adjustment purposes. If provided, it will be summed instead of the
number of instance (Default: NULL)}

\item{input.summary.dt}{A data.table with DAOH summarised by value (and by
some other group if desired).}
}
\value{
data.table with totals and proportions per group of DAOH values
}
\description{
Plots a bar plot using ggplot2. Input to this function must be summarised, it
will not count for you.
}
