% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert.date.cols.auto}
\alias{convert.date.cols.auto}
\title{Tries to detect date columns by looking at the first entry. Doesn't work as such.}
\usage{
convert.date.cols.auto(input.dt, date.col.names = NULL)
}
\arguments{
\item{input.dt}{The DAOH data with date columns}

\item{date.col.names}{The columns to convert.}
}
\description{
Tries to detect date columns by looking at the first entry. Doesn't work as such.
}
