% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert.date.cols.fasttime}
\alias{convert.date.cols.fasttime}
\title{Convert dates of relevant columns by reference, so as not to copy big-as
tables (i.e. doesn't return anything, modifies the input.) This one uses
fasttime, which is a lot faster (...) if you've got big big data. However,
times need to be in a specific format.}
\usage{
convert.date.cols.fasttime(input.dt, date.cols)
}
\arguments{
\item{input.dt}{The DAOH data with date columns}

\item{date.cols}{The columns to convert.}
}
\description{
Convert dates of relevant columns by reference, so as not to copy big-as
tables (i.e. doesn't return anything, modifies the input.) This one uses
fasttime, which is a lot faster (...) if you've got big big data. However,
times need to be in a specific format.
}
