% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert.date.cols}
\alias{convert.date.cols}
\title{Convert dates of relevant columns by reference, so as not to copy big-as
tables (i.e. doesn't return anything, modifies the input.)}
\usage{
convert.date.cols(input.dt, date.cols, fmat = "\%Y-\%m-\%d")
}
\arguments{
\item{input.dt}{The DAOH data with date columns}

\item{date.cols}{The columns to convert.}
}
\description{
Convert dates of relevant columns by reference, so as not to copy big-as
tables (i.e. doesn't return anything, modifies the input.)
}
