% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{detect.overlapping.events}
\alias{detect.overlapping.events}
\title{Merge overlapping admissions so that days are not counted twice.}
\usage{
detect.overlapping.events(
  daoh.event.dt,
  patient.id.col.name = "PRIM_HCU.IDencrypted",
  index.event.id.col.name = "index.event.id",
  admission.start.col.name = "EVSTDATE",
  admission.end.col.name = "EVENDATE",
  daoh.period.start.col.name = "daoh.period.start",
  daoh.period.end.col.name = "daoh.period.end",
  merge.event.id.col.name = "merge.event.id"
)
}
\arguments{
\item{daoh.event.dt}{A data.table with patient identifier, index event
identifier, and all admission found within the DAOH period of that index
event, generated by prune.events.to.period occurred. Will be modified by
reference to have a new column grouping overlapping sets of events.}

\item{patient.id.col.name}{Character name of the patient identifier column
(Default: 'PRIM_HCU.IDencrypted')}

\item{index.event.id.col.name}{Character name of the index event identifier
column (Default: 'index.event.id')}

\item{admission.start.col.name}{Character name of the column containing
admission start dates (Default: 'EVSTDATE')}

\item{admission.end.col.name}{Character name of the column containing
admission end dates (Default: 'EVENDATE')}

\item{daoh.period.start.col.name}{Character name of the column that will be
generated in index.op.dt to contain the start of each DAOH period (Default:
'daoh.period.start')}

\item{daoh.period.end.col.name}{Character name of the column that will be
generated in index.op.dt to contain the end of each DAOH period (Default:
'daoh.period.end')}

\item{merge.event.id.col.name}{Character name of the column that will be
populated with unique identifiers of overlapping events (Default:
'merge.event.id')}
}
\value{
The modified input daoh.event.dt data.table, with a new column with a
  unique identifier for each overlapping group of admissions (in column
  merge.event.id.col.name). The input will be modified by reference, so it's
  not strictly necessary to use the output.
}
\description{
Merge overlapping admissions so that days are not counted twice.
}
