% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate.all.site.tsds}
\alias{generate.all.site.tsds}
\title{Generate timed sampling distributions for all sites in a}
\usage{
generate.all.site.tsds(
  input.dt,
  site.dt,
  cluster.dt,
  n.sampling.period,
  outcome.col.name,
  values = 0:89,
  smooth.func = smooth.daoh,
  output.to.console = F
)
}
\arguments{
\item{input.dt}{Data for all sites as data.table}

\item{site.dt}{Data.table about sites and their clusters.}

\item{cluster.dt}{Data.table about clusters and when they start/stop}

\item{n.sampling.period}{How many periods should the distribution be divided
into?}

\item{outcome.col.name}{Name of outcome column.}

\item{values}{Values that timed sampling dist can take.}

\item{smooth.func}{How to smooth the distribution.}

\item{output.to.console}{Output progress to console.}

\item{sampling.period.out.start}{Time to start the timed sampling dist}

\item{sampling.period.out.length}{Length of each period in timed sampling
dist}
}
\value{
The timed sampling distribution (note however that it is set in the
  site by reference)
}
\description{
Generate timed sampling distributions for all sites in a
}
