% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.adjust.R
\name{generate.dra.riskgp.columns}
\alias{generate.dra.riskgp.columns}
\title{Generates risk group columns.}
\usage{
generate.dra.riskgp.columns(
  input.dt,
  model.list,
  riskgp.col.name = "riskgp",
  n.groups = 10
)
}
\arguments{
\item{input.dt}{Input data.table, will be modified by reference.}

\item{model.list}{Models to risk adjust on, columns will be given the suffix
of the name of the corresponding list value. Will be coerced to list if not
provided (which may mess with naming, e.g. providing a single would result
in a list with no name).}

\item{riskgp.col.name}{What should the risk group column be called? (And what
should component columns start with?). (default: 'riskgp').}

\item{n.groups}{Number of groups to generate per risk factor (default: 10)}
}
\value{
The modified input.dt with length(model.list)+1 extra columns,
  unnecessary to use it as it will be modified by reference. Column
  riskgp.col.name will be a factor, designed to use in the generate.weights
  function.
}
\description{
Generates risk group columns from a list of models and inserts them as
reference. Component columns for each model will named with a common prefix,
followed by the name of that value in model.list, and an overall risk group.
}
