% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.adjust.R
\name{generate.quantreg.model}
\alias{generate.quantreg.model}
\title{Generate a quantile regression model on which to risk adjust.}
\usage{
generate.quantreg.model(
  input.dt,
  output.variable,
  covariates,
  jitter.output.variable = FALSE,
  quantiles.to.assess = seq(0.25, 0.75, by = 0.1)
)
}
\arguments{
\item{input.dt}{Data.table with output.variable and covariates.}

\item{output.variable}{Character name of outcome to be adjusted.}

\item{covariates}{Character vector of column names to adjust upon.}

\item{jitter.output.variable}{controls whether the output.variable will be
jittered. It seems like DAOH is not quite continuous enough for quantreg.
Perhaps not inherent to DAOH per se, but related to the distribution of the
CheckWHO data (Default: FALSE).}

\item{quantiles.to.assess}{Numeric vector, values between 0 and 1, of models
to fit. Output variable will be adjusted by the model associated with the
nearest quantile.}
}
\value{
A quantile regression model fitted at requested quantiles.
}
\description{
Generate a quantile regression model on which to risk adjust.
}
