% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.adjust.R
\name{generate.riskgp.vector}
\alias{generate.riskgp.vector}
\title{Functions for generating the data needed for adjustment.}
\usage{
generate.riskgp.vector(
  input.df,
  model,
  n.groups,
  riskgp.level.name.suffix = ""
)
}
\arguments{
\item{input.df}{A data.frame with the columns from the regression model.}

\item{model}{The regression model, should be compatible with base::predict}

\item{n.groups}{Number of groups to generate.}

\item{riskgp.level.name.suffix}{Suffix to place after the number of the risk
group (default: '')}
}
\value{
A vector of risk groups in the same order as the rows of input.df.
  Higher risk groups are more likely to score highly in the model, and vice
  versa. Groups will be names e.g. '01' + riskgp.level.name.suffix
}
\description{
Creates a number of risk groups from a regression model. Higher risk groups
are more likely to score highly in the model, and vice versa.
}
