% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate.site.tsd}
\alias{generate.site.tsd}
\title{Generate the timed sampling distributions from real data.}
\usage{
generate.site.tsd(this.site.dt, n.sampling.period, outcome.col.name,
  sampling.period.out.start, sampling.period.out.length, values = 0:89,
  smooth.func = smooth.daoh, site = NULL)
}
\arguments{
\item{this.site.dt}{The data as a data.table, with time and outcome columns.}

\item{n.sampling.period}{How many periods should the distribution be divided
into?}

\item{outcome.col.name}{Name of outcome column.}

\item{sampling.period.out.start}{Time to start the timed sampling dist}

\item{sampling.period.out.length}{Length of each period in timed sampling
dist}

\item{values}{Values that timed sampling dist can take.}

\item{smooth.func}{How to smooth the distribution.}

\item{site}{The SW.Site object to set by reference}
}
\value{
The timed sampling distribution (note however that it is set in the
  site by reference if site is not NULL)
}
\description{
Will also set the patients per unit time.
}
