% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate.smoothed.sampling.dist}
\alias{generate.smoothed.sampling.dist}
\title{Plot DAOH in a histogram}
\usage{
generate.smoothed.sampling.dist(input.dist, values = NULL,
  smooth.func = smooth.daoh)
}
\arguments{
\item{input.dist}{The DAOH dataset on which to base the sampling
distribution, as a vector.}

\item{values}{The value that can the sampling distribution can take (default
min(input.dist):max(input.dist), which will fail if not integer input)}
}
\value{
ggplot2 plot
}
\description{
Plot DAOH in a histogram
}
