% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate.timed.sampling.dist}
\alias{generate.timed.sampling.dist}
\title{Generate a timed.sampling.dist object from an input distribution (with
timings obvs)}
\usage{
generate.timed.sampling.dist(input.dt, outcome.col.name,
  n.sampling.period = 1, sampling.period.in.start = NULL,
  sampling.period.in.length = NULL, sampling.period.out.start = NULL,
  sampling.period.out.length = NULL, smooth.func = smooth.daoh,
  values = NULL)
}
\arguments{
\item{input.dt}{The (data.table) input distribution to emulate with outcome
and time columns}

\item{outcome.col.name}{The name of the outcome column}

\item{n.sampling.period}{How many sampling periods?}

\item{sampling.period.in.start}{At what time should the sampling start?
Should be integer.}

\item{sampling.period.in.length}{How long should each time period be in the
output? Should be integer.}

\item{sampling.period.out.start}{At what time should the probability
distribution in the output start? Should be integer.}

\item{smooth.func}{What function should be used for smoothing? Mainly made
for DAOH, so that's default. Should be integer.}

\item{values}{The value that can the sampling distribution can take (default
min(input.dist):max(input.dist), which will fail if not integer input)}

\item{sampling.period.out.start}{How long should each period be in the
output? Should be integer.}
}
\value{
A Timed.Sampling.Dist object
}
\description{
Generate a timed.sampling.dist object from an input distribution (with
timings obvs)
}
