% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{modify.daoh.distribution}
\alias{modify.daoh.distribution}
\title{Modify a DAOH distribution, until at.quantile is target.quantile.val.
Uses data.table so that things can be done by reference.}
\usage{
modify.daoh.distribution(x = 0:90, prob = rep(1, 90),
  target.quantile.val = NULL, relative.quantile.val = NULL,
  at.quantile = 0.5, multiple.per.step = 0.01, output.to.console = T,
  empirical.dist = F, n.emp = 1e+05)
}
\arguments{
\item{x}{Possible values for daoh}

\item{prob}{Probability of each daoh value}

\item{target.quantile.val}{The target value for at.quantile}

\item{relative.quantile.val}{Value that the target quantile should be relative to its current value.}

\item{at.quantile}{The quantile at which to insert a difference (default: 0.5).}

\item{multiple.per.step}{Probabilities will be adjusted as a function of this number.}

\item{output.to.console}{Print progress to console.}

\item{empirical.dist}{Calculate quantiles for calculating cost empirically (slower)}

\item{n.emp}{Number of samples that will be generated when empirically finding quantiles}
}
\value{
The modified distribution as a list containing values and weights.
}
\description{
Modify a DAOH distribution, until at.quantile is target.quantile.val.
Uses data.table so that things can be done by reference.
}
