% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{modify.timed.daoh.distribution}
\alias{modify.timed.daoh.distribution}
\title{Modify a timed DAOH distribution.}
\usage{
modify.timed.daoh.distribution(tsd, mod.time.begin, mod.time.end,
  target.quantile.val = NULL, relative.quantile.val = NULL, ...)
}
\arguments{
\item{tsd}{The Timed.Sampling.Distribution object.}

\item{mod.time.begin}{The time to start the modification.}

\item{mod.time.end}{The time at which to end the modification.}

\item{target.quantile.val}{The target value for the quantile of interest (by
default the median)}

\item{...}{Arguments to be sent to modify.daoh.distribution}
}
\description{
If there is no split at time.begin and/or time.end, one will be inserted. Each
period between mod.time.begin and mod.time.end will then be individually
modified.
}
