% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.daoh.barplot}
\alias{plot.daoh.barplot}
\title{Plot DAOH barplot.}
\usage{
\method{plot}{daoh.barplot}(
  input.summary.dt,
  daoh.col.name = "daoh",
  by.group = NULL,
  xlimits = c(-0.5, 90.5),
  y.aes = c("prop", "N")[1]
)
}
\arguments{
\item{input.summary.dt}{A data.table with DAOH summarised by value (and by
some other group if desired).}

\item{daoh.col.name}{Name of the DAOH column}

\item{by.group}{Name of the stratifying group column, none if NULL (e.g., an
intervention) (Default: NULL)}

\item{xlimits}{Limits on the x-axis of the plot (Default: c(-0.5, 90.5))}

\item{y.aes}{What's on the y-axis. "prop" for proportion per group, "N" for
frequency. (Default:  c('prop', "N")[1])}
}
\value{
A ggplot2 plot.
}
\description{
Plots a bar plot using ggplot2. Input to this function must be summarised, it
will not count for you.
}
