% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.daoh.density}
\alias{plot.daoh.density}
\title{Plot DAOH in a density chart}
\usage{
\method{plot}{daoh.density}(input.dt, by.group = NA_character_, xlimits = c(0, 90), adjust = 1.5)
}
\arguments{
\item{input.dt}{The DAOH data (with daoh column)}

\item{by.group}{The group (string) by which to separate the histograms.}
}
\value{
ggplot2 plot
}
\description{
Plot DAOH in a density chart
}
