% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.distribution}
\alias{plot.distribution}
\title{Plot the distribution by drawing from it heaps.}
\usage{
\method{plot}{distribution}(x, prob, n = 1e+08)
}
\arguments{
\item{x}{The values in the distribution}

\item{prob}{The probability that the value will be drawn.}

\item{n}{The number of points to sample. It actually has to be pretty high
for DAOH, about 1e8 (10,000,000)}
}
\value{
ggplot2 plot
}
\description{
Plot the distribution by drawing from it heaps.
}
