% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{prune.events.to.period}
\alias{prune.events.to.period}
\title{Match index events to admissions that fall within the corresponding DAOH
period.}
\usage{
prune.events.to.period(
  index.op.dt,
  event.dt,
  daoh.limits,
  patient.id.col.name = "PRIM_HCU.IDencrypted",
  index.event.id.col.name = "index.event.id",
  admission.start.col.name = "EVSTDATE",
  admission.end.col.name = "EVENDATE",
  daoh.period.start.col.name = "daoh.period.start",
  daoh.period.end.col.name = "daoh.period.end"
)
}
\arguments{
\item{index.op.dt}{A data.table that minimally has a patient identifier, a
unique index event identifier for each event, and a date on which it
occurred. Will be modified by reference to have a DAOH start and end date.}

\item{event.dt}{A data.table of all other admission events that occurred,
minimally with a patient identifier and event start and end dates.}

\item{daoh.limits}{Numeric vector of length two with DAOH start and end
relative to the index event.}

\item{patient.id.col.name}{Character name of the patient identifier column
(Default: 'PRIM_HCU.IDencrypted')}

\item{index.event.id.col.name}{Character name of the index event identifier
column (Default: 'index.event.id')}

\item{admission.start.col.name}{Character name of the column containing
admission start dates (Default: 'EVSTDATE')}

\item{admission.end.col.name}{Character name of the column containing
admission end dates (Default: 'EVENDATE')}

\item{daoh.period.start.col.name}{Character name of the column that will be
generated in index.op.dt to contain the start of each DAOH period (Default:
'daoh.period.start')}

\item{daoh.period.end.col.name}{Character name of the column that will be
generated in index.op.dt to contain the end of each DAOH period (Default:
'daoh.period.end')}
}
\value{
A data.table with patient identifier, index event identifier, and all
  admission found within the DAOH period of that index event.
}
\description{
Match index events to admissions that fall within the corresponding DAOH
period.
}
