% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{quantile.dist.emp}
\alias{quantile.dist.emp}
\title{Empirically a quantile from a sampling distribution.}
\usage{
\method{quantile}{dist.emp}(x, prob, quantile = 0.5, n = 1e+06)
}
\arguments{
\item{x}{The values.}

\item{prob}{The weights (same length as values).}

\item{quantile}{The quantile to find.}

\item{n}{The number of random values to generate.}
}
\value{
Value at requested quantile.
}
\description{
Empirically a quantile from a sampling distribution.
}
