% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{quantile.dist}
\alias{quantile.dist}
\title{Find a quantile from a sampling distribution.}
\usage{
\method{quantile}{dist}(x, prob, quantile = 0.5)
}
\arguments{
\item{x}{The values.}

\item{prob}{The weights (same length as values).}

\item{quantile}{The quantile to find.}
}
\value{
Value at requested quantile.
}
\description{
Find a quantile from a sampling distribution.
}
