% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.adjust.R
\name{risk.adjust.on.quantreg.model}
\alias{risk.adjust.on.quantreg.model}
\title{Apply the quantile regression models to risk adjust a data.table.}
\usage{
risk.adjust.on.quantreg.model(input.dt, qr.model, by.reference = FALSE)
}
\arguments{
\item{input.dt}{Data.table with output.variable and covariates.}

\item{qr.model}{Quantreg model generated by generate.quantreg.model}

\item{by.reference}{Logical, whether to modify the input data.table by
inserting the risk adjusted variable, and any dummy variables (Default:
FALSE).}
}
\description{
Apply the quantile regression models to risk adjust a data.table.
}
