% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{smooth.daoh}
\alias{smooth.daoh}
\title{A smoothing function particularly for DAOH.}
\usage{
smooth.daoh(input.freq.dt)
}
\arguments{
\item{input.freq.dt}{Data.table with columns 'outcome' and 'N'}
}
\value{
Data.table with same columns as input, but also 'probs' for feeding
  into sample()
}
\description{
A smoothing function particularly for DAOH.
}
