% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{std.err}
\alias{std.err}
\title{Calculate standard error, with or without NA}
\usage{
std.err(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Remove NA first}
}
\value{
Standard error
}
\description{
Calculate standard error, with or without NA
}
