% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.daoh.histogram}
\alias{plot.daoh.histogram}
\title{Plot DAOH in a histogram}
\usage{
\method{plot}{daoh.histogram}(input.dt, by.group = NA_character_)
}
\arguments{
\item{input.dt}{The DAOH data (with daoh column)}

\item{by.group}{The group (string) by which to separate the histograms.}
}
\value{
ggplot2 plot
}
\description{
Plot DAOH in a histogram
}
