% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.daoh.histogram}
\alias{plot.daoh.histogram}
\title{Plot DAOH in a histogram}
\usage{
\method{plot}{daoh.histogram}(
  input.dt,
  daoh.col.name = "daoh",
  by.group = NULL,
  xlimits = c(-0.5, 90.5),
  y.aes = c("prop", "N")[1]
)
}
\arguments{
\item{input.dt}{The DAOH data (with daoh column)}

\item{by.group}{The group (string) by which to separate the histograms.}

\item{xlimits}{Length two numeric vector of xlimits..}

\item{y.aes}{What's on the y-axis. "prop" for proportion per group, "N" for
frequency. (Default:  c('prop', "N")[1])}
}
\value{
ggplot2 plot
}
\description{
Plot DAOH in a histogram
}
